% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-params.R
\name{bandwidth}
\alias{bandwidth}
\title{Parametre dials pour la bande passante spatiale}
\usage{
bandwidth(range = c(10L, 300L), trans = NULL)
}
\arguments{
\item{range}{Bornes entieres de la bande passante.}

\item{trans}{Transformation optionnelle.}
}
\value{
Un parametre \code{dials}.
}
\description{
Bande passante utilisee par les estimateurs locaux de type GWR/MGWR/MGWRSAR.
}
