% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-diagnose-spatial.R
\name{binary_auc_value}
\alias{binary_auc_value}
\title{Aire sous la courbe ROC (Mann-Whitney U), sans dependance a yardstick}
\usage{
binary_auc_value(truth01, prob)
}
\description{
Aire sous la courbe ROC (Mann-Whitney U), sans dependance a yardstick
}
\keyword{internal}
