% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51-parsnip-inlaspde.R
\name{inla_spde_default_mesh}
\alias{inla_spde_default_mesh}
\title{Maillage SPDE par defaut, derive des donnees}
\usage{
inla_spde_default_mesh(coords, max_edge = NULL, cutoff = NULL, offset = NULL)
}
\description{
Memes considerations que \code{inla_spde_default_priors()}: parametres
\code{fmesher::fm_mesh_2d_inla()} calcules depuis la diagonale de la bounding
box des coordonnees, sans reglage manuel par jeu de donnees.
}
\keyword{internal}
