% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51-parsnip-inlaspde.R
\name{inlaspde_pred_impl}
\alias{inlaspde_pred_impl}
\title{Fonction interne de prediction du champ SPDE INLA pour parsnip}
\usage{
inlaspde_pred_impl(object, new_data)
}
\description{
Fonction interne de prediction du champ SPDE INLA pour parsnip
}
\keyword{internal}
