% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-params.R
\name{k_neighbors}
\alias{k_neighbors}
\title{Parametre dials pour le nombre de voisins kNN}
\usage{
k_neighbors(range = c(3L, 30L), trans = NULL)
}
\arguments{
\item{range}{Bornes entieres du nombre de voisins.}

\item{trans}{Transformation optionnelle.}
}
\value{
Un parametre \code{dials}.
}
\description{
Nombre de voisins utilise pour construire la matrice de poids spatiaux W.
}
