% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-parsnip-mgwrsar.R
\name{mgwrsar_pred_impl}
\alias{mgwrsar_pred_impl}
\title{Fonction interne de prediction mgwrsar pour parsnip}
\usage{
mgwrsar_pred_impl(object, new_data, coords)
}
\description{
Fonction interne de prediction mgwrsar pour parsnip
}
\keyword{internal}
