% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-params.R
\name{mstop}
\alias{mstop}
\title{Parametre dials pour mstop}
\usage{
mstop(range = c(50L, 1500L), trans = NULL)
}
\arguments{
\item{range}{Bornes entieres du nombre d'iterations.}

\item{trans}{Transformation optionnelle.}
}
\value{
Un parametre \code{dials}.
}
\description{
Nombre d'iterations de boosting utilise par \code{spboost}.
}
