% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-parsnip-probitspatial.R
\name{probitspatial_fit_impl}
\alias{probitspatial_fit_impl}
\title{Fonction interne de fit du probit spatial pour parsnip}
\usage{
probitspatial_fit_impl(
  formula,
  data,
  coords,
  W = NULL,
  model_type = "SAR",
  k_neighbors = 8,
  style = "W",
  zero_policy = TRUE
)
}
\description{
Fonction interne de fit du probit spatial pour parsnip
}
\keyword{internal}
