% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-parsnip-probitspatial.R
\name{probitspatial_pred_impl}
\alias{probitspatial_pred_impl}
\title{Fonction interne de prediction du probit spatial pour parsnip}
\usage{
probitspatial_pred_impl(object, new_data)
}
\description{
Retourne toujours la probabilite predite (classe positive = second niveau
stocke au fit); \code{set_pred()} construit les sorties "class"/"prob" a partir
de ce vecteur, comme le fait \code{parsnip} lui-meme pour \code{logistic_reg()}.
}
\keyword{internal}
