% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-parsnip-spatialreg.R
\name{spatialreg_fit_impl}
\alias{spatialreg_fit_impl}
\title{Fonction interne de fit spatialreg pour parsnip}
\usage{
spatialreg_fit_impl(
  formula,
  data,
  coords,
  W = NULL,
  model_type = "SAR",
  k_neighbors = 8,
  style = "W",
  zero_policy = TRUE,
  pred_type = "TS"
)
}
\description{
Fonction interne de fit spatialreg pour parsnip
}
\keyword{internal}
