% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-parsnip-spatialreg.R
\name{spatialreg_pred_impl}
\alias{spatialreg_pred_impl}
\title{Fonction interne de prediction spatialreg pour parsnip}
\usage{
spatialreg_pred_impl(object, new_data)
}
\description{
Fonction interne de prediction spatialreg pour parsnip
}
\keyword{internal}
