% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-parsnip-spboost.R
\name{spboost_fit_impl}
\alias{spboost_fit_impl}
\title{Fonction interne de fit spboost pour parsnip}
\usage{
spboost_fit_impl(
  formula,
  data,
  coords,
  DGP = "SAR",
  method = NULL,
  mstop = 500L,
  nu = 0.1,
  k_neighbors = 8
)
}
\description{
Fonction interne de fit spboost pour parsnip
}
\keyword{internal}
