% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-parsnip-spboost.R
\name{spboost_pred_impl}
\alias{spboost_pred_impl}
\title{Fonction interne de prediction spboost pour parsnip}
\usage{
spboost_pred_impl(object, new_data)
}
\description{
Fonction interne de prediction spboost pour parsnip
}
\keyword{internal}
