% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-params.R
\name{spmoran_enum}
\alias{spmoran_enum}
\title{dials parameter for the number of Moran eigenvectors}
\usage{
spmoran_enum(range = c(5L, 200L), trans = NULL)
}
\arguments{
\item{range}{Integer bounds for the number of eigenvectors.}

\item{trans}{Optional transformation.}
}
\value{
A \code{dials} parameter.
}
\description{
Number of Moran eigenvectors used by \code{spmoran::meigen_f()}. Small values
keep tuning fast; larger values allow richer spatial filtering bases.
}
