% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-parsnip-spmoran.R
\name{spmoran_pred_impl}
\alias{spmoran_pred_impl}
\title{Internal spmoran prediction function for parsnip}
\usage{
spmoran_pred_impl(object, new_data)
}
\description{
Internal spmoran prediction function for parsnip
}
\keyword{internal}
