% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-estimator-registration.R
\name{unregister_spatial_estimator}
\alias{unregister_spatial_estimator}
\title{Remove a registered custom spatial estimator}
\usage{
unregister_spatial_estimator(id)
}
\arguments{
\item{id}{Estimator name previously passed to \code{\link[=register_spatial_estimator]{register_spatial_estimator()}}.}
}
\value{
\code{id}, invisibly.
}
\description{
Remove a registered custom spatial estimator
}
