% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-spatial-weights.R
\name{build_knn_listw}
\alias{build_knn_listw}
\title{Construire un objet listw kNN}
\usage{
build_knn_listw(coords, k = 8)
}
\arguments{
\item{coords}{Matrice ou data frame de coordonnees.}

\item{k}{Nombre de voisins.}
}
\value{
Un objet \code{listw}.
}
\description{
Construit la version \code{spdep::listw} de la meme structure de voisinage que
\code{build_knn_W()}. Cette forme est utilisee par \code{spatialreg}.
}
