% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_benchmark_datasets}
\alias{spatial_benchmark_datasets}
\alias{georgia}
\alias{columbus_crime}
\alias{london_hp}
\alias{boston_housing}
\alias{dub_voter}
\alias{ewhp}
\alias{lasrosas}
\title{Registered spatial benchmark datasets}
\format{
Data frames or \code{sf} objects with prepared coordinate columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 159 rows and 18 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 49 rows and 22 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 316 rows and 24 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 506 rows and 40 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 322 rows and 14 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 519 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3443 rows and 14 columns.
}
\usage{
georgia

columbus_crime

london_hp

boston_housing

dub_voter

ewhp

lasrosas
}
\description{
Prepared spatial regression datasets bundled with \code{spatialtidymodels}.
Each object can be loaded with \verb{data(<name>, package = "spatialtidymodels")}
and is also known by \code{benchmark_spatial_dataset()}.
}
\section{Available objects}{

\code{georgia}, \code{columbus_crime}, \code{london_hp}, \code{boston_housing}, \code{dub_voter},
\code{ewhp}, and \code{lasrosas}.
}

\examples{
data(columbus_crime, package = "spatialtidymodels")
head(columbus_crime)

}
\keyword{datasets}
