% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-benchmark-spatial.R
\name{spatial_dataset_spec}
\alias{spatial_dataset_spec}
\title{Define a dataset for \code{benchmark_spatial_datasets()}}
\usage{
spatial_dataset_spec(name, data, formula, coords)
}
\arguments{
\item{name}{Short dataset name.}

\item{data}{Data frame.}

\item{formula}{Model formula.}

\item{coords}{Coordinate column names.}
}
\value{
A \code{spatial_dataset_spec} object.
}
\description{
Define a dataset for \code{benchmark_spatial_datasets()}
}
