% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-benchmark-spatial.R
\name{spatial_dataset_spec}
\alias{spatial_dataset_spec}
\title{Definir un jeu de donnees pour \code{benchmark_spatial_datasets()}}
\usage{
spatial_dataset_spec(name, data, formula, coords)
}
\arguments{
\item{name}{Nom court du dataset.}

\item{data}{Donnees.}

\item{formula}{Formule econometrique.}

\item{coords}{Colonnes de coordonnees.}
}
\value{
Une liste de classe \code{spatial_dataset_spec}.
}
\description{
Definir un jeu de donnees pour \code{benchmark_spatial_datasets()}
}
