% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-parsnip-spatialreg.R
\name{spatialreg_fit_impl}
\alias{spatialreg_fit_impl}
\title{Fonction interne de fit spatialreg pour parsnip}
\usage{
spatialreg_fit_impl(formula, data, coords, model_type = "SAR", k_neighbors = 8)
}
\description{
Fonction interne de fit spatialreg pour parsnip
}
\keyword{internal}
