% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialtidymodels-package.R
\docType{package}
\name{spatialtidymodels-package}
\alias{spatialtidymodels}
\alias{spatialtidymodels-package}
\title{spatialtidymodels}
\description{
Extension tidymodels en developpement pour des estimateurs de regression
spatiale.
}
\details{
Le package enregistre des specs \code{parsnip} et des moteurs pour les backends
\code{spatialreg}, \code{spboost} et \code{mgwrsar}. Les routes \code{spatialreg} disposent de
tests de parite numerique contre le benchmark manuel sur plusieurs datasets
du projet.
}
\author{
\strong{Maintainer}: Johnny D'Oliveira \email{johnny@example.com}

}
\keyword{internal}
