% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-parsnip-spmoran.R
\name{spmoran_esf_reg}
\alias{spmoran_esf_reg}
\title{parsnip specification for ESF}
\usage{
spmoran_esf_reg(mode = "regression", coords = NULL, vif = 10, enum = NULL)
}
\arguments{
\item{mode}{parsnip mode. Only \code{"regression"} is supported.}

\item{coords}{Character vector of length 2 giving the coordinate columns.}

\item{vif}{Variance inflation threshold passed to \code{spmoran::esf()}.}

\item{enum}{Optional number of Moran eigenvectors for \code{spmoran::meigen_f()}.}
}
\value{
A \code{parsnip} model specification.
}
\description{
Convenience constructor for \code{spmoran::esf()}.
}
