% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-parsnip-spmoran.R
\name{spmoran_resf_reg}
\alias{spmoran_resf_reg}
\title{parsnip specification for RE-ESF}
\usage{
spmoran_resf_reg(mode = "regression", coords = NULL, enum = NULL)
}
\arguments{
\item{mode}{parsnip mode. Only \code{"regression"} is supported.}

\item{coords}{Character vector of length 2 giving the coordinate columns.}

\item{enum}{Optional number of Moran eigenvectors for \code{spmoran::meigen_f()}.}
}
\value{
A \code{parsnip} model specification.
}
\description{
Convenience constructor for \code{spmoran::resf()}.
}
